feat: record counts in map after POST /count
The endpoint is now POST /count. We no longer include the date as part
of the endpoint, because the date is now encoded within the ID of the
counter element.

These IDs are stored in a map. When compiling the totals for each
date, shared dates will be accumulated into one figure, e.g.

06-04-1 10
06-05-2 4
06-04-3 12

will become

06-04 22
06-05 4